requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Computer Vision
opencv-python>=4.8.0
ultralytics>=8.0.0
albumentations>=1.3.0
pillow>=9.5.0

# Tracking
filterpy>=1.4.5
scikit-image>=0.21.0
scipy>=1.10.0

# API & Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
streamlit>=1.25.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-contrib-python>=4.8.0

# MLOps
mlflow>=2.5.0
wandb>=0.15.0

# Database
sqlalchemy>=2.0.0
pymongo>=4.4.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
pyyaml>=6.0
python-json-logger>=2.0.7

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Deployment
gunicorn>=21.2.0

# AWS (Optional)
boto3>=1.28.0
